+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Improved Sync Assistant
+- Trim fields in email config
 
 ## [0.8.32] - 2022-06-07
 ### Added: